Remove sumUpSplits function & correct originalTransaction method
