+++ src/components/ReportActionItem/TransactionPreview/index.tsx
@@ -1,7 +1,6 @@
 import {useRoute} from '@react-navigation/native';
 import React, {useCallback, useMemo} from 'react';
 import type {GestureResponderEvent} from 'react-native';
-import type {OnyxEntry} from 'react-native-onyx';
 import PressableWithoutFeedback from '@components/Pressable/PressableWithoutFeedback';
 import {showContextMenuForReport} from '@components/ShowContextMenuContext';
 import useLocalize from '@hooks/useLocalize';
@@ -12,8 +11,8 @@ import {convertToDisplayString} from '@libs/CurrencyUtils';
 import {canUseTouchScreen} from '@libs/DeviceCapabilities';
 import {getOriginalMessage, isMoneyRequestAction as isMoneyRequestActionReportActionsUtils} from '@libs/ReportActionsUtils';
 import {getTransactionDetails} from '@libs/ReportUtils';
-import {getReviewNavigationRoute} from '@libs/TransactionPreviewUtils';
-import {getTransaction, isCardTransaction, removeSettledAndApprovedTransactions} from '@libs/TransactionUtils';
+import {getOriginalTransactionIfBillIsSplit, getReviewNavigationRoute} from '@libs/TransactionPreviewUtils';
+import {isCardTransaction, removeSettledAndApprovedTransactions} from '@libs/TransactionUtils';
 import Navigation from '@navigation/Navigation';
 import type {PlatformStackRouteProp} from '@navigation/PlatformStackNavigation/types';
 import type {TransactionDuplicateNavigatorParamList} from '@navigation/types';
@@ -22,27 +21,9 @@ import {clearIOUError} from '@userActions/Report';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type SCREENS from '@src/SCREENS';
-import type {Transaction} from '@src/types/onyx';
 import {isEmptyObject} from '@src/types/utils/EmptyObject';
 import TransactionPreviewContent from './TransactionPreviewContent';
 import type {TransactionPreviewProps} from './types';
-
-const getOriginalTransactionIfBillIsSplit = (transaction: OnyxEntry<Transaction>) => {
-    const {originalTransactionID, source, splits} = transaction?.comment ?? {};
-
-    // If splits property is defined in the transaction, it is actually an original transaction
-    if (splits && splits.length > 0) {
-        return {isBillSplit: true, originalTransaction: transaction};
-    }
-
-    if (!originalTransactionID || source !== CONST.IOU.TYPE.SPLIT) {
-        return {isBillSplit: false, originalTransaction: transaction};
-    }
-
-    const originalTransaction = getTransaction(originalTransactionID);
-
-    return {isBillSplit: !!originalTransaction, originalTransaction: originalTransaction ?? transaction};
-};
 
 function TransactionPreview(props: TransactionPreviewProps) {
     const {translate} = useLocalize();
